--- keywords/admin/__init__.py
@@ -1,1 +1,0 @@
-from .keyword import KeywordAdmin
