+++ product/admin/__init__.py
@@ -1,0 +1,6 @@
+from .category import CategoryAdmin
+from .code import CodeAdmin
+from .product import ProductAdmin
+from .product_code import ProductCodeAdmin
+from .product_feedback import ProductFeedbackAdmin
+from .product_tag import ProductTagAdmin
