+++ product/admin/product.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
-from products.models import Product
+from product.models import Product
 
 @admin.register(Product)
 class ProductAdmin(admin.ModelAdmin):
