+++ product/models/__init__.py
@@ -1,0 +1,6 @@
+from .product import Product
+from .code import Code
+from .category import Category
+from .product_code import ProductCode
+from .product_feedback import ProductFeedback
+from .product_tag import ProductTag
