+++ product/models/product.py
@@ -4,7 +4,7 @@ from django.db import models
 
 class Product(models.Model):
     id = models.AutoField(primary_key=True, null=False, blank=False) # primary key
-    category_id = models.ForeignKey('products.Category', null = False, blank = False, on_delete = models.CASCADE)
+    category_id = models.ForeignKey('product.Category', null = False, blank = False, on_delete = models.CASCADE)
     name = models.CharField(max_length = 400, null = False, blank = False)
     producer = models.CharField(max_length = 400, null = False, blank = False)
     feedback_cnt = models.IntegerField()
