+++ product/models/product_code.py
@@ -1,5 +1,5 @@
 from django.db import models
-from products.models import Product, Code
+from product.models import Product, Code
 
 class ProductCode(models.Model):
     id = models.AutoField(primary_key=True, null=False, blank=False)
