+++ product/models/product_tag.py
@@ -1,8 +1,8 @@
 from django.db import models
-from products.models import Product
-from keywords.models import Keyword
+from product.models import Product
+from tag.models import Tag
 
-class ProductKeyword(models.Model):
+class ProductTag(models.Model):
     id = models.AutoField(primary_key=True, null=False, blank=False)
     product_id = models.ForeignKey(Product, null = False, blank = False, on_delete = models.CASCADE)
-    keyword_id = models.ForeignKey(Keyword, on_delete = models.CASCADE)
+    tag_id = models.ForeignKey("tag.Tag", on_delete = models.CASCADE)
