--- products/admin/__init__.py
@@ -1,6 +1,0 @@
-from .category import CategoryAdmin
-from .code import CodeAdmin
-from .product import ProductAdmin
-from .productcode import ProductCodeAdmin
-from .productfeedback import ProductFeedbackAdmin
-from .productkeyword import ProductKeywordAdmin
