--- products/models/__init__.py
@@ -1,6 +1,0 @@
-from .product import Product
-from .code import Code
-from .category import Category
-from .productcode import ProductCode
-from .productfeedback import ProductFeedback
-from .productkeyword import ProductKeyword
